add: retire support for <v1 contracts
