+++ src/lib/wallet/GoodWallet.js
@@ -1,6 +1,5 @@
 import Config from '../../config/config'
 import { GoodWallet } from './GoodWalletClass'
-import './PaymentLinks'
 export default new GoodWallet({
   web3Transport: Config.web3TransportProvider,
 })
